+++ tests/additionals/test_functions.py
@@ -1,10 +1,13 @@
+import pytest
+
 from wetterdienst.additionals.functions import check_parameters, retrieve_time_resolution_from_filename, \
     retrieve_parameter_from_filename, retrieve_period_type_from_filename, determine_parameters, \
-    create_station_data_dtype_mapping, cast_to_list
+    create_station_data_dtype_mapping, cast_to_list, parse_enumeration_from_template
 from wetterdienst.enumerations.column_names_enumeration import DWDMetaColumns, DWDDataColumns
 from wetterdienst.enumerations.period_type_enumeration import PeriodType
 from wetterdienst.enumerations.time_resolution_enumeration import TimeResolution
 from wetterdienst.enumerations.parameter_enumeration import Parameter
+from wetterdienst.exceptions.invalid_parameter_exception import InvalidParameter
 
 
 def test_check_parameters():
@@ -59,3 +62,11 @@ def test_cast_to_list():
     assert cast_to_list("abc") == ["abc"]
     assert cast_to_list(1) == [1]
     assert cast_to_list(PeriodType.HISTORICAL) == [PeriodType.HISTORICAL]
+
+
+def test_parse_enumeration_from_template():
+    assert parse_enumeration_from_template("climate_summary", Parameter) == Parameter.CLIMATE_SUMMARY
+    assert parse_enumeration_from_template("kl", Parameter) == Parameter.CLIMATE_SUMMARY
+
+    with pytest.raises(InvalidParameter):
+        parse_enumeration_from_template("climate", Parameter)