+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - change updating "parallel" argument to be done after parameter parsing to prevent mistakenly not found 
 parameter
 - remove find_all_match_strings function and extract functionality to individual operations
+- parameter, time resolution and period type can now also be passed as strings of the enumerations e.g.
+"climate_summary" or "CLIMATE_SUMMARY" for Parameter.CLIMATE_SUMMARY
 
 ## 0.1.0
 - initial release